Skip recipient if it is an unknown instead of fail
When user doesn't exists anymore on an smtp server we get:

```
v1.0.18 - Error during execute - Error setting to:: 550 5.1.1 <an.unkownuser@orange.com>: Recipient address rejected: User unknown
```

And the resource will stop sending. But it is more interesting to only notified that this recipient will be skipped instead of triggering error to continu sending.

This PR add this behaviour instead of the current behaviour.